Add main class and missing raw experiment data
