+++ .gitignore
@@ -109,3 +109,5 @@ fabric.properties
 hs_err_pid*
 
 # End of https://www.gitignore.io/api/kotlin,intellij
+
+.DS_Store
