+++ simResults/varying reqsPerSec/reqresults.csv
@@ -1,0 +1,101 @@
+reqPerEdgePerSecond;min_latency;avg_latency;max_latency
+100;15;32,3367;102
+200;15;37,0156;118
+300;15;40,3768;119
+400;15;42,3969;113
+500;15;43,8634;127
+600;15;45,3684;128
+700;15;47,2978;127
+800;15;49,7982;129
+900;15;52,5543;130
+1000;15;55,1928;132
+1100;15;57,6086;134
+1200;15;59,8706;133
+1300;15;61,8955;133
+1400;15;63,6451;131
+1500;15;65,2482;134
+1600;15;66,6497;134
+1700;15;67,9388;135
+1800;15;69,1104;134
+1900;15;70,0502;135
+2000;15;70,9964;134
+2100;15;71,8710;134
+2200;15;72,6509;135
+2300;15;73,4253;134
+2400;15;74,1444;134
+2500;15;74,7199;135
+2600;15;75,3278;134
+2700;15;75,8467;134
+2800;15;76,3200;134
+2900;15;76,7444;133
+3000;15;77,2220;134
+3100;15;77,6100;134
+3200;15;78,0000;134
+3300;15;78,3505;135
+3400;15;78,7012;135
+3500;15;79,0117;134
+3600;15;79,3329;135
+3700;15;79,5898;134
+3800;15;79,8658;134
+3900;15;80,1450;134
+4000;15;80,4009;135
+4100;15;80,6481;135
+4200;15;80,8155;134
+4300;15;81,0162;134
+4400;15;81,2384;135
+4500;15;81,4309;134
+4600;15;81,6313;134
+4700;15;81,8093;135
+4800;15;82,0137;134
+4900;15;82,1478;134
+5000;15;82,3022;135
+5100;15;82,4752;135
+5200;15;82,6300;134
+5300;15;82,7928;134
+5400;15;82,9414;134
+5500;15;83,0748;134
+5600;15;83,2117;134
+5700;15;83,2949;135
+5800;15;83,3994;135
+5900;15;83,5125;134
+6000;15;83,6270;135
+6100;15;83,7658;134
+6200;15;83,8779;134
+6300;15;83,9401;134
+6400;15;84,0271;134
+6500;15;84,1207;135
+6600;15;84,2595;135
+6700;15;84,3196;135
+6800;15;84,3888;135
+6900;15;84,4404;135
+7000;15;84,5454;135
+7100;15;84,6740;135
+7200;15;84,7397;135
+7300;15;84,8312;134
+7400;15;84,8841;135
+7500;15;84,9436;134
+7600;15;85,0087;135
+7700;15;85,0481;135
+7800;15;85,1190;135
+7900;15;85,1858;135
+8000;15;85,2961;135
+8100;15;85,3716;135
+8200;15;85,4126;135
+8300;15;85,4845;134
+8400;15;85,5257;134
+8500;15;85,5792;134
+8600;15;85,6155;134
+8700;15;85,6547;135
+8800;15;85,6916;135
+8900;15;85,7391;135
+9000;15;85,8518;135
+9100;15;85,8932;134
+9200;15;85,9358;135
+9300;15;85,9480;135
+9400;15;85,9934;135
+9500;15;86,0485;135
+9600;15;86,0831;135
+9700;15;86,1343;135
+9800;15;86,1563;135
+9900;15;86,1928;135
+10000;15;86,2336;135
